requirements.txt
# Core Dependencies
numpy>=1.24.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Deep Learning - GPU Support
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8 with GPU support

# Tracking
filterpy>=1.4.5
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0

# Data Processing
pandas>=2.0.0
scikit-learn>=1.3.0
openpyxl>=3.1.0

# Video Processing
moviepy>=1.0.3

# Utils
tqdm>=4.65.0
pyyaml>=6.0
